make nav more modular
