+++ dipole/templates/index.html
@@ -11,7 +11,7 @@
     <script>
       var username = "{{ username }}";
     </script>
-    <script type="module" crossorigin src="/static/assets/index-j6ZgLBIX.js"></script>
+    <script type="module" crossorigin src="/static/assets/index-DcB7bPOm.js"></script>
     <link rel="stylesheet" crossorigin href="/static/assets/index-CuLNt5s6.css">
   </head>
   <body>
